+++ src/fractal/burning_ship.c
@@ -6,18 +6,18 @@
 /*   By: edarnand <edarnand@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/28 15:31:25 by edarnand          #+#    #+#             */
-/*   Updated: 2025/03/03 11:20:50 by edarnand         ###   ########.fr       */
+/*   Updated: 2025/03/04 13:46:56 by edarnand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fractol.h"
 
-static int	calc_burning_ship(double c_real, double c_im, int max_iteration)
+static double	calc_burning_ship(double c_real, double c_im, int max_iteration)
 {
 	double	real;
 	double	im;
 	double	tmp_im;
-	int		index;
+	double	index;
 
 	real = 0;
 	im = 0;
@@ -32,7 +32,7 @@ static int	calc_burning_ship(double c_real, double c_im, int max_iteration)
 	return (index);
 }
 
-int	burning_ship(t_complex *comp, int max_iteration)
+double	burning_ship(t_complex *comp, int max_iteration)
 {
 	return (calc_burning_ship(comp->real_curr, comp->im_curr, max_iteration));
 }
